+++ Source/OxyPlot.Wpf/PlotView.cs
@@ -128,7 +128,7 @@ namespace OxyPlot.Wpf
         {
             var scale = base.UpdateDpi();
             this.RenderContext.DpiScale = scale;
-            this.RenderContext.VisualOffset = this.TransformToAncestor(this.GetAncestorWindowFromVisualTree(this)).Transform(default);
+            this.RenderContext.VisualOffset = this.TransformToAncestor(this.GetAncestorVisualFromVisualTree(this)).Transform(default);
             return scale;
         }
 
@@ -146,18 +146,21 @@ namespace OxyPlot.Wpf
 
 
         /// <summary>
-        /// Returns a reference to the window object that hosts the dependency object in the visual tree.
+        /// Returns a reference to the visual object that hosts the dependency object in the visual tree.
         /// </summary>
         /// <returns> The host window from the visual tree.</returns>
-        private Window GetAncestorWindowFromVisualTree(DependencyObject startElement)
+        private Visual GetAncestorVisualFromVisualTree(DependencyObject startElement)
         {
-            DependencyObject parent = startElement;
-            while (!(parent is Window))
+            
+            DependencyObject child = startElement;
+            DependencyObject parent = VisualTreeHelper.GetParent(child);
+            while (parent != null)
             {
-                if (parent == null) { break; }
-                parent = VisualTreeHelper.GetParent(parent);
+                child = parent;
+                parent = VisualTreeHelper.GetParent(child);
             }
-            return parent as Window ?? Window.GetWindow(this);
+
+            return child is Visual visualChild ? visualChild : Window.GetWindow(this);
         }
     }
 }